Batch rule for MD5 checks
We use awk to add the local path to the fastq files, which allows us to run md5sum from the main working directory.
GNU parallel runs one line of the file at a time (-N1) and prints the output to the log, so it is not erased in case there are some files that fail the checksum.
The token .done file is only created if all files pass the checksum test, and is required by downstream processes as a safety check.