Trim first 11 bp to avoid adapter bias
This is done via seqtk, invoked as a wildcard rule that depends on rule all to specify desired output files.
I also include the results of multiqc as an extra input as a safety measure: if the report is missing or not updated, it will require that it is re-compiled so we can always check that the trimmed base pairs make sense with a correct version of the QC report.