Compile summary of contrast coefficients
The final output is a list of data.frames, each of which contains the coefficient for a specific gene in each treatment.
NOTE: The coefficients of LPS treatments are interaction terms between time and treatment. On order to get the true total LFC for those we need to first add the coefficients for time