Add volcano plot and output for GSEA analysis
This version of the script implements the ad-hoc version of the linear model but also returns the standard format output for downstream GSEA processing: a list of gene IDs ranked by fold change expression. Each element in the list is one different contrast of interest.

We return 2 lists, one for downregulation enrichment and one for upregulation enrichment. They are the same, but one is sorted on -logFC instead of FC